+++ internal/meta/label.go
@@ -1,7 +1,7 @@
 package meta
 
 import (
-	"github.com/giantswarm/apiextensions/v3/pkg/label"
+	"github.com/giantswarm/k8smetadata/pkg/label"
 
 	"github.com/giantswarm/konfigure/pkg/project"
 )